api/green/requirements.txt
Django==1.8.6
djangorestframework==3.3.1
Markdown==2.6.4
mysqlclient==1.3.7
wheel==0.24.0
